+++ tests/HttpClientRequestTest.php
@@ -122,4 +122,67 @@ class HttpClientRequestTest extends TestCase
 
         self::assertSame('Baz', $request->getRawContent());
     }
+
+    /**
+     * Test getCACertificate method.
+     */
+    public function testGetCACertificate()
+    {
+        $request = new HttpClientRequest(Url::parse('https://example.com/foo/bar'));
+
+        self::assertNull($request->getCACertificate());
+    }
+
+    /**
+     * Test setCACertificate method.
+     */
+    public function testSetCACertificate()
+    {
+        $request = new HttpClientRequest(Url::parse('https://example.com/foo/bar'));
+        $request->setCACertificate(FilePath::parse('/foo.pem'));
+
+        self::assertSame(DIRECTORY_SEPARATOR . 'foo.pem', $request->getCACertificate()->__toString());
+    }
+
+    /**
+     * Test getClientCertificate method.
+     */
+    public function testGetClientCertificate()
+    {
+        $request = new HttpClientRequest(Url::parse('https://example.com/foo/bar'));
+
+        self::assertNull($request->getClientCertificate());
+    }
+
+    /**
+     * Test setClientCertificate method.
+     */
+    public function testSetClientCertificate()
+    {
+        $request = new HttpClientRequest(Url::parse('https://example.com/foo/bar'));
+        $request->setClientCertificate(FilePath::parse('/bar.pem'));
+
+        self::assertSame(DIRECTORY_SEPARATOR . 'bar.pem', $request->getClientCertificate()->__toString());
+    }
+
+    /**
+     * Test getClientKey method.
+     */
+    public function testGetClientKey()
+    {
+        $request = new HttpClientRequest(Url::parse('https://example.com/foo/bar'));
+
+        self::assertNull($request->getClientKey());
+    }
+
+    /**
+     * Test setClientKey method.
+     */
+    public function testSetClientKey()
+    {
+        $request = new HttpClientRequest(Url::parse('https://example.com/foo/bar'));
+        $request->setClientKey(FilePath::parse('/baz.pem'));
+
+        self::assertSame(DIRECTORY_SEPARATOR . 'baz.pem', $request->getClientKey()->__toString());
+    }
 }