+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## Unreleased
+- Added getCACertificate and setCACertificate methods to HttpClientRequest.
+- Added getClientCertificate and setClientCertificate methods to HttpClientRequest.
+- Added getClientKey and setClientKey methods to HttpClientRequest.
 
 ## 1.0.0 - 2019-01-24
 - First stable revision.